% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lapse_rate.R
\name{fun_lapse_rate}
\alias{fun_lapse_rate}
\title{Saturated vapour pressure}
\usage{
fun_lapse_rate(DATA_FOR_LAPSERATE, ELEVATION_VECTOR)
}
\arguments{
\item{DATA_FOR_LAPSERATE}{data.frame of data from stations}

\item{ELEVATION_VECTOR}{vector which contain the elevations of stations in the same order as in DATA_FOR_LAPSERATE}

\item{VARIABLE_FOR_LAPSERATE}{character of variable name which it is apply the lapse rate}
}
\description{
Saturated vapour pressure
}
