% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_all_stations.R
\name{fun_read_all_stations}
\alias{fun_read_all_stations}
\title{Function that read from a folder all file inside and extract the variable selected}
\usage{
fun_read_all_stations(FOLDER, FILES, VARIABLE)
}
\arguments{
\item{FOLDER}{package folder (where package is cloned)}

\item{FILES}{vector of files available in folder data/Input/. Files must have a single raw of headers and all file must start and end at the same time}

\item{VARIABLE}{variable to extract from every single file and merge in one single dataframe}
}
\description{
Function that read from a folder all file inside and extract the variable selected
}
