% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dew_point.R
\name{SatVapPressure}
\alias{SatVapPressure}
\title{Saturated vapour pressure}
\usage{
SatVapPressure(TEMP, P)
}
\arguments{
\item{TEMP}{Air Temperature}

\item{P}{Air pressure}
}
\description{
Saturated vapour pressure
}
