% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dew_point.R
\name{Tdew}
\alias{Tdew}
\title{Dew Point Temperature}
\usage{
Tdew(TEMP, RH, Z)
}
\arguments{
\item{TEMP}{Air temperature}

\item{RH}{Relative Humidity}

\item{Z}{elevation}
}
\description{
Dew Point Temperature
}
