% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dew_point.R
\name{TfromSatVapPressure}
\alias{TfromSatVapPressure}
\alias{TfromSatVapPressure}
\title{Temperature from saturated water vapour}
\usage{
TfromSatVapPressure(e, P)

TfromSatVapPressure(e, P)
}
\arguments{
\item{e}{Saturated vapour pressure}

\item{P}{Air pressure}

\item{SVP}{Saturated vapour pressure}

\item{P}{Air pressure}
}
\description{
Temperature from saturated water vapour

Temperature from saturated water vapour
}
